arch: serial: open file outputs for append
When opening the same output file for two serial devices (earlycon and
console), the output would be overwritten by the later device.

Change the file creation to use append mode so that the output file
contains all of the logs from both devices instead of overwriting it.

Tested with:

  crosvm run \
  --serial hardware=serial,type=file,console=false,earlycon=true,path=test.log \
  --serial hardware=virtio-console,type=file,console=true,stdin=true,path=test.log \
  vm_kernel

BUG=None
TEST=see above - verify log contains earlycon and console output

Change-Id: I14dab9eaf56dbb0ae410215324b20b34fea723ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2208712
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>