balloon: Use different method for sizing balloon.
Use free + cached memory in the guest instead of available
reported through the balloon stats, which uses some strange heuristics,
for sizing the balloon. This brings it more in line to what
the ChromeOS available counter does.

Also, change the way the memory is balanced such that the
balloon is sized so that the "available" memory in the guest
would be equal the host's available memory.

BUG=b:160826299
TEST=arc.MemoryChromeOSPerf.vm tast test improves to be closer to
 container ARC: 1900MB -> 2400MB on 4GB machine.
 No difference in arc.MemoryAndroidPerf.vm.

Change-Id: Idddf1e5f5d75ea4d1ebb190b60dd785bbe2327df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2299279
Auto-Submit: Suleiman Souhlal <suleiman@chromium.org>
Tested-by: Suleiman Souhlal <suleiman@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Suleiman Souhlal <suleiman@chromium.org>