linux: replace sort with sort_by_key and Reverse
Use the _by_key variant of the sort function as recommended by
https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_sort_by
and replace the potentially easy-to-miss swap of the two compared values
with an explicit call to std::cmp::Reverse (also recommended by the same
clippy lint).

Fixes a clippy warning with Rust 1.46.0.

BUG=None
TEST=docker/wrapped_smoke_test.sh (with Rust 1.46.0)

Change-Id: I682652736a27cb09111dbc813e654b17269670da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2386479
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>