devices: irqchip: register events in IOAPIC
Previously the caller of ioapic::new was responsible for calling
register_irq_event on the IOAPIC's outgoing events.  The MSI IrqRoutes
for these outgoing events used the same GSIs < 24 as the incoming
PIC/IOAPIC routes.

This worked for the split irqchip, because PIC/IOAPIC events are handled
in userspace and MSI events in the kernel, so there was no confusion
about which route an event should use.  But with the userspace irqchip,
the routes have to be differentiated, otherwise incoming and outgoing
IOAPIC events with the same GSI trigger each other in a loop.  Also,
registering all outgoing IOAPIC events upfront increased the number of
irq event tokens by 24, even though only a few outgoing events are used.

Now the IOAPIC is responsible for registering outgoing events and it
allocates new GSIs >= 24 for its MSI routes.  An event and route are
only set up when a redirection table entry is configured with non-zero
MSI data.

BUG=chromium:1077058
TES=boots with split irqchip

Change-Id: I87c1df6150d50205810759ef152e1c9b33bcfe4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2419672
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Colin Downs-Razouk <colindr@google.com>