devices: gpu: enable the shader disk cache with new mesa default
A new build option for Mesa allows it to be built with support
for the disk cache but with it disabled by default. The
MESA_GLSL_CACHE_DISABLE environment variable must be set to
false to enable it in that case.

BUG=b:168540438
TEST=run with --gpu=cache-path=/tmp,cache-size=50M and
     confirm that files are created in /tmp/mesa_shader_cache.

Change-Id: I456e224ec4d2d8ed1f20689706ff4752b5706bb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2485901
Tested-by: John Bates <jbates@chromium.org>
Auto-Submit: John Bates <jbates@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: John Bates <jbates@chromium.org>